requirements.txt
streamlit==1.28.1
langchain==0.1.13
langchain-openai==0.1.1
langchain-community==0.0.29
openai==1.12.0
python-dotenv==1.0.1
pandas==2.0.3
sqlalchemy==2.0.23
httpx==0.27.2